requirements.txt
attrs==21.2.0
autopep8==1.6.0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
Flask==2.0.2
Flask-WTF==1.0.0
idna==3.3
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
packaging==21.2
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
PyHamcrest==2.0.2
pyparsing==2.4.7
pytest==6.2.5
python-dotenv==0.19.2
requests==2.26.0
toml==0.10.2
urllib3==1.26.7
Werkzeug==2.0.2
WTForms==3.0.0
